Extract common functions to separate module
Closes: https://github.com/HowProgrammingWorks/NodejsStarterKit/issues/117
