+++ lib/auth.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { crypto } = require('./dependencies.js');
+const { crypto, common } = require('./dependencies.js');
 
 const BYTE  = 256;
 const TOKEN = 'token';
@@ -42,12 +42,6 @@ const parseCookies = cookie => {
     values[key] = val.trim();
   }
   return values;
-};
-
-const parseHost = host => {
-  const portOffset = host.indexOf(':');
-  if (portOffset > -1) return host.substr(0, portOffset);
-  return host;
 };
 
 module.exports = application => {
@@ -95,7 +89,7 @@ module.exports = application => {
 
   const start = (client, userId) => {
     const token = generateToken();
-    const host = parseHost(client.req.headers.host);
+    const host = common.parseHost(client.req.headers.host);
     const ip = client.req.connection.remoteAddress;
     const cookie = `${TOKEN}=${token}; ${COOKIE_HOST}=${host}; HttpOnly`;
     const sandbox = getSandbox();
@@ -132,7 +126,7 @@ module.exports = application => {
   };
 
   const remove = (client, token) => {
-    const host = parseHost(client.req.headers.host);
+    const host = common.parseHost(client.req.headers.host);
     client.res.setHeader('Set-Cookie', COOKIE_DELETE + host);
     sessions.delete(token);
     db.delete('Session', { token });
