+++ lib/dependencies.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const api = {};
+const api = { common: require('./common.js') };
 const internals = [
   'util', 'child_process', 'worker_threads', 'os', 'v8', 'vm', 'path', 'url',
   'assert', 'querystring', 'string_decoder', 'perf_hooks', 'async_hooks',
