+++ lib/server.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { http, https, path, worker } = require('./dependencies.js');
+const { http, https, path, worker, common } = require('./dependencies.js');
 
 const WebSocket = require('ws');
 const Semaphore = require('./semaphore.js');
@@ -20,10 +20,6 @@ const LONG_RESPONSE = 30000;
 const METHOD_OFFSET = '/api/'.length;
 
 const clients = new Map();
-
-const timeout = msec => new Promise(resolve => {
-  setTimeout(resolve, msec);
-});
 
 const receiveArgs = async req => new Promise(resolve => {
   const body = [];
@@ -154,11 +150,9 @@ const listener = application => (req, res) => {
     });
   } else {
     if (url === '/' && !req.connection.encrypted) {
-      const { host } = req.headers;
-      const portOffset = host.indexOf(':');
-      const addr = portOffset > -1 ? host.substr(0, portOffset) : host;
-      const port = application.server.randomPort();
-      client.redirect(`https://${addr}:${port}/`);
+      const host = common.parseHost(req.headers.host);
+      const port = common.sample(application.server.ports);
+      client.redirect(`https://${host}:${port}/`);
     }
     client.static();
   }
@@ -187,15 +181,11 @@ class Server {
     this.instance.listen(port, host);
   }
 
-  randomPort() {
-    return this.ports[Math.floor(Math.random() * this.ports.length)];
-  }
-
   async close() {
     this.instance.close(err => {
       if (err) this.application.logger.error(err.stack);
     });
-    await timeout(SHUTDOWN_TIMEOUT);
+    await common.timeout(SHUTDOWN_TIMEOUT);
     closeClients();
   }
 }
